requirements.txt
appnope==0.1.2
asgiref==3.4.0
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
decorator==5.0.9
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
Faker==8.8.2
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
ipython==7.25.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mixer==7.1.2
mysqlclient==2.0.3
numpy==1.21.1
openapi-codec==1.3.2
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pluggy==0.13.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pytesseract==0.3.8
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.14
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
